Updates Rails integration for CLI tools
When running in a Rails application the postqueue CLI tool expects
the application to configure its postqueue setting as part of the
application, for example by an initializer in config/initializers/postqueue.rb.

When not running inside a Rails application the postqueue CLI tool expects
a config/postqueue.rb file to contain postqueue configuration.